Issues  Related to HashSet is resolved.
Class name and File should be same, so that issue is resolved.
Placed appropriate print Statements for better Understanding.
